SAK-43357 allow rubrics component to save evaluations
https://jira.sakaiproject.org/browse/SAK-43357
