+++ assignment/tool/src/java/org/sakaiproject/assignment/tool/AssignmentToolUtils.java
@@ -422,26 +422,6 @@ public class AssignmentToolUtils {
                 alerts.addAll(integrateGradebook(options, aReference, associateGradebookAssignment, null, null, null, -1, null, sReference, "remove", -1));
             }
 
-            String siteId = (String) options.get("siteId");
-
-            Map<String, String> rubricsParams
-                = options.entrySet().stream().filter(e -> e.getKey().startsWith("rbcs"))
-                    .collect(Collectors.toMap(e -> e.getKey(), e -> (String) e.getValue()));
-
-            // Persist the rubric evaluations
-            if (!rubricsParams.isEmpty()) {
-                if (rubricsService.hasAssociatedRubric(RubricsConstants.RBCS_TOOL_ASSIGNMENT, submission.getAssignment().getId(), siteId)){
-
-                    if (StringUtils.isNotBlank(siteId)) {
-                        rubricsParams.put("siteId", siteId);
-                    }
-                    for (AssignmentSubmissionSubmitter submitter : submission.getSubmitters()) {
-                        String submitterId = submitter.getSubmitter();
-                        rubricsService.saveRubricEvaluation(RubricsConstants.RBCS_TOOL_ASSIGNMENT, submission.getAssignment().getId(), submission.getId(), submitterId, submission.getGradedBy(), rubricsParams);
-                    }
-                }
-            }
-
             try {
                 return assignmentService.getSubmission(submission.getId());
             } catch (Exception e) {
