+++ samigo/samigo-app/src/webapp/jsf/evaluation/rubricModal.jsp
@@ -8,9 +8,6 @@
 			entity-id='<h:outputText value="pub.#{questionScores.publishedId}.#{questionScores.itemId}"/>'
 			evaluated-item-id='<h:outputText value="#{description.assessmentGradingId}.#{questionScores.itemId}" />'
 			item-id='<h:outputText value="#{description.assessmentGradingId}"/>'
-			<h:panelGroup rendered="#{description.rubricStateDetails != ''}">
-				state-details='<h:outputText value="#{description.rubricStateDetails}"/>'
-			</h:panelGroup>>
 		</sakai-rubric-grading>
 	</div>
 </h:panelGroup>
